+++ Buyers.txt
@@ -1,0 +1,4 @@
+Mihnea Velcea has just bought 12 tickets.
+
+Mihnea Velcea has just bought 12 tickets.
+
