+++ CMakeLists.txt
@@ -98,7 +98,9 @@ FetchContent_MakeAvailable(googletest)
 ###############################################################################
 
 # NOTE: update executable name in .github/workflows/cmake.yml:25 when changing name here
-add_executable(${PROJECT_NAME} main.cpp ${SOURCES})
+add_executable(${PROJECT_NAME} main.cpp ${SOURCES}
+        include/TakeAnAction.h
+        include/CreateTickets.h)
 
 ###############################################################################
