--- include/example.h
@@ -1,4 +1,0 @@
-#ifndef OOP_EXAMPLE_H
-#define OOP_EXAMPLE_H
-int doSomething(int x);
-#endif //OOP_EXAMPLE_H
