+++ src/Codecopy.h
@@ -1,0 +1,14 @@
+//
+// Created by Mihnea on 3/22/2024.
+//
+
+#ifndef OOP_CODECOPY_H
+#define OOP_CODECOPY_H
+
+
+class Codecopy {
+
+};
+
+
+#endif //OOP_CODECOPY_H
