--- src/example.cpp
@@ -1,5 +1,0 @@
-#include "example.h"
-
-int doSomething(int x) {
-    return x;
-}
